services/inference/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
mlflow==2.14.2
psycopg2-binary==2.9.9
pydantic==2.10.4
numpy==1.26.4
prometheus-client==0.21.1
boto3==1.34.162
